Move /const & /utils from /services to /src, rename /const to /data
