+++ src/mock/tasks.js
@@ -1,6 +1,6 @@
-import getRandomInt from "../services/utils/getRandomInt";
-import DAYS from "../services/const/days";
-import COLORS from "../services/const/colors";
+import getRandomInt from "../utils/getRandomInt";
+import DAYS from "../data/days";
+import COLORS from "../data/colors";
 
 const MockDescriptions = [
   `Изучить теорию`,
