+++ src/views/menu.js
@@ -1,4 +1,4 @@
-import createElement from "../services/utils/createElement";
+import createElement from "../utils/createElement";
 
 const createMenuItemTemplate = (item) => {
   const {id, title, isChecked} = item;
