+++ src/views/task.js
@@ -1,4 +1,4 @@
-import createElement from "../services/utils/createElement";
+import createElement from "../utils/createElement";
 
 const createTagsTemplate = (tags) => {
   return Array.from(tags)
